+++ tests/spec/jsAxios/schema.js
@@ -17,14 +17,14 @@ export var ContentType;
   ContentType["UrlEncoded"] = "application/x-www-form-urlencoded";
 })(ContentType || (ContentType = {}));
 export class HttpClient {
-  constructor({ securityWorker, ...axiosConfig } = {}) {
+  constructor({ securityWorker, secure, ...axiosConfig } = {}) {
     this.securityData = null;
     this.setSecurityData = (data) => {
       this.securityData = data;
     };
     this.request = async ({ secure, path, type, query, format = "json", body, ...params }) => {
       const secureParams =
-        ((typeof secure === "boolean" ? secure : this.baseApiParams.secure) &&
+        ((typeof secure === "boolean" ? secure : this.secure) &&
           this.securityWorker &&
           (await this.securityWorker(this.securityData))) ||
         {};
@@ -42,6 +42,7 @@ export class HttpClient {
       });
     };
     this.instance = axios.create({ ...axiosConfig, baseURL: axiosConfig.baseURL || "https://api.github.com" });
+    this.secure = secure;
     this.securityWorker = securityWorker;
   }
   mergeRequestParams(params1, params2) {